refactored SimdFFT namespace and related project structure
