+++ algorithm/test/CMakeLists.txt
@@ -3,14 +3,13 @@ add_ut_test(qa_DataSetEstimators)
 add_ut_test(qa_FilterTool)
 add_ut_test(qa_ImChart)
 add_ut_test(qa_SchmittTrigger)
-add_ut_test(qa_simd_fft)
+add_ut_test(qa_SimdFFT)
 target_link_libraries(qa_algorithm_fourier PRIVATE gnuradio-algorithm)
 target_link_libraries(qa_DataSetEstimators PRIVATE gnuradio-algorithm)
 target_link_libraries(qa_FilterTool PRIVATE gnuradio-algorithm)
 target_link_libraries(qa_ImChart PRIVATE gnuradio-algorithm)
 target_link_libraries(qa_SchmittTrigger PRIVATE gnuradio-algorithm)
+target_link_libraries(qa_SimdFFT PRIVATE gnuradio-algorithm)
 
 add_executable(example_ImChart example_ImChart.cpp)
 target_link_libraries(example_ImChart PRIVATE gnuradio-algorithm)
-
-add_subdirectory(incubator)
